agent-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
google-genai>=1.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
